Add Remove Book Functionality
Wrote function to add event listener to remove button which deletes
book from library array as well as removed it from the DOM. Linked the
book card to the library by assigning the array index as a data attribute.
Got stuck on the fact needed to assign it to every button each time using
querySelectorAll and then forEach.