Feat: custom form validation
Add custom error messages using setCustomValidity when adding a new book.
Had to work around setCustomValidity function not preventing default
messages if the inputs were left untouched and the form was submitted.
To fix this, needed to explicitly call setCustomValidity for all fields
when the modal opens (i.e., during form initialization) and ensure the
fields are properly validated at that time.